clio.pullrequest: Break cycles in the requested merge set before assessing merges.
When handling many pull requests at once from multiple users, it's possible that a cycles exist in the graph of requested merges, which violates the code's assumption that 'root' bodies can be found.
We now break the cycles at an arbitrary edge.  The resulting mergeset will still have all the correct bodies, but the merge direction will be decided in the coercion step.